third_party/WebKit/LayoutTests/fast/dom/SelectorAPI/id-fastpath-expected.txt

This tests that the querySelector and querySelectorAll fast path for IDs is not overzelous.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.querySelector('div#testid') is correctNode
PASS document.querySelector('#testid') is correctNode
PASS document.querySelector('ul#testid') is null
PASS document.querySelector('ul #testid') is null
PASS document.querySelector('#testid[attr]') is null
PASS document.querySelector('#testid:not(div)') is null
PASS document.querySelectorAll('div#testid').length is 1
PASS document.querySelectorAll('div#testid').item(0) is correctNode
PASS document.querySelectorAll('#testid').length is 1
PASS document.querySelectorAll('#testid').item(0) is correctNode
PASS document.querySelectorAll('ul#testid').length is 0
PASS document.querySelectorAll('ul #testid').length is 0
PASS document.querySelectorAll('#testid[attr]').length is 0
PASS document.querySelectorAll('#testid:not(div)').length is 0
PASS successfullyParsed is true

TEST COMPLETE
